Change bash prompt
Instead of using the tip of the current path, use the current time. I
don't generally need the current path in every command, and can print
`pwd` if I need to remember.

I might not need the time either, but it is useful if I've left the
terminal alone for a while and can't remember when I ran a command last.