Roll back notmuch usage
Instead of using notmuch queries and virtual mailboxes, roll back to
using neomutt's native IMAP/Maildir support. This seems more responsive
in neomutt, and is less confusing for me, and maybe most importantly, I
can get it to work fully.

Use replace neomutt's search with a notmuch query mechanism, which is to
me much superior to neomutt's built in search.

Adjust the mail sync script to do notmuch tagging that matches the
Maildir format exactly, with the one exception of tagging anything with
the keyword "unsubscribe" anywhere in the message as part of a list. Add
one virtual mailbox for mails tagged as lists.

Make an improvement to shell script handling; set `wait_key` to false,
meaning that when neomutt calls an external script, it won't make me
press enter after the command executes.