Stop being fancy with EDITOR/VISUAL variables
Setting these variables to `nvim -u NORC` is a cool idea, in that it
reduces startup time but in practice has some problems.

The first is that not all commands understand an EDITOR variable that
contains flags. For example, `crontab -e` gets pretty confused and bails
out.

Another problem is that I don't actually like the experience that much.
I get confused since the editor is different when I'm inside an already
running neovim instance or not, and it increases my cognitive load just
enough to be annoying.