Make mailsync tagging more concise
Use notmuch's batch syntax to shorten the code for email tagging on
sync.

Remove the lists tagging, and remove from that virtual mailbox from
neomutt. I wasn't really using it and I am using a different method to
handling newsgroups.

Switch to a more concise way of setting the Trash.