Fixup VISUAL environment variable
The VISUAL and EDITOR environment variable is getting clobbered inside
neovim; when running in terminal mode in neovim, I want VISUAL to be
`nvr`, and outside neovim I want it to be `nvim`.

Wrap the bash VISUAL export to make sure it doesn't get called inside a
running neovim instance.