Neovim config still works if no plugins present
There are only a few places outside custom ftplugins and after files that operate on plugins being available.

One, is adding fugitive data to the statusline, and another is the colorscheme.

Guard the fugitive statusline addition in a fugitive function detect.

Suppress error messages if the colorscheme isn't in the path. The colorscheme will still load if it's present on the system. Otherwise, we go back to the default theme.

No fuss, no muss.