Replace vim-grepper with built in vim :grep
After using the straight up :grep command for a bit I found it more than adequate for my needs, and I can more easily pass it arbitrary arguments.

Configure :grep to use ripgrep if it's present on the system, and remove vim-grepper.

TODO: One thing I miss is the operator function from vim-grepper, but that shouldn't be too hard to replicate in my own code.