Change git message recovery strategy
Trying to load in the built-in git backup of the message doesn't always
work, in that the message file can be empty in my experience, even after
I *have* filled out a message.

Instead, try saving the contents of the git message buffer into the Vim
g register. If the commit fails, I can try again without fear of
overriding that register, since I never use g. If I need to restart Vim,
I can always save it to a file.