Remove documentation plugins
Remove the PHP and MDN documentation plugins. Having that stuff locally
is cool, but it's ultimately a hack for now, and feels like it.

The PHP docs clutter up the disk and can sometimes get confused for Vim
help, and the formatting is done via scraping, so it's not always up to
stuff.

The MDN docs are slow to query, and suffer the same formatting issues as
PHP, both being scraped. It's often easier to just open the web MDN
page.

Remove in the anticipation of writing some `K` wrappers for
PHP/JS/CSS/HTML to open the appropriate page on the web.