Add some basic documentation settings
Use `dict` to lookup words for markdown files. Since the output is
usually pretty long, open it in a new split. The output is also similar
to man pages, so running `Man!` over it provides some nice formatting.

Use PHP's built in function signature documentation. Since the output is
short, keep it in the command line window.

TODO: Open the PHP function signatures in the preview window or
something like that.