Fix fugitive mappings not loading
My vim-fugitive mappings load in an after plugin hook, that checks if
the `:Git` command is present, which fugitive supplies, before doing any
work. This used to work, but fugitive has since become more efficient,
only initializing this command if a file in a git directory is loaded.

To compensate, change the check to look for the fugitive plugin on disk;
it will still load the mappings when there is no fugitive present - aka
opening a non-git tracked file - but this is okay since they're pretty
out of the way and it would be more overhead to keep checking if
fugitive is loaded on every file open.