Git plugin re-working
Remove the HG plugin, since I haven't worked with that in a while.

Add a mergetool improvement plugin - this has worked out nicely in tests
for me, as I try to move more diff/merging outside external dependencies
and into vim.

Add a cool floating window plugin to display git commit info. This could
be a really useful, lightweight replacement for a lot of what I use
fugitive blame for.