Swap editorconfig with undotree
Editorconfig is fairly slow, and I do not switch projects all that much. It only takes me a bit of time to configure vim to match the project settings, so it is not needed all the time. Undotree is really useful, on the other hand, and I think it is time I brought it back.