Remove -f option from fugitive git mergetool
The `-f` option in the example I was looking at is to force Vim to the
foreground, however, this is only an option for regular vim/nvim
binaries, not the neovim-remote binary I'm calling, where it doesn't
make sense (a neovim instance is already running in this scenario).

Hence, it was just creating a new empty buffer called -f, which isn't
what anyone wants.

Remove.