Add bash support for fish
Enable this by using the fisher fish package manager. This allows the
use of nvm in fish.

Also, add a new method of reverse search. It's only okay, but works
better than the hacked together solution I half made (in that it works).
I wish you could see a few more results of your search, not just one,
but it's still an improvement.