Remove nvi support
Let's be honest, I don't use vi anymore. It was really fun and useful to
use it and see what is baked into the most minimal version, and while
I'm not a full "dark mode" vim user anymore, I'm not fully "light"
either. Somewhere in the middle. And that requires Neovim (for me).