Remove vim twig plugin
The twig plugin is unmaintained, and it's not perfect. The HTML Django
syntax is also not perfect, since it's a different language, but it's
close enough and built into vim. It also has compatibility with
vim-endwise, which is really useful. Makes sense to lighten up the
configuration and get the added functionality in exchange for not quite
as good syntax highlighting.