Use php-support fork of vim-closer
The vim-closer plugin has a PR opened against it that provides PHP
support for closing braces. I use PHP every day, so this is really
useful for me. The PR has been opened for a while, and despite my
urging, has not been merged. Seems like the project is not super
maintained. Since this is really useful, point to the contributor's PR
to get PHP support; the upstream repo is not ahead of the contributor's
and given it's slow development, will probably stay that way if/when
that PR is merged.