Fix EDITOR variable inside neovim session
There's an order of operations problem with setting the EDITOR and
VISUAL environment variables inside fish configuration. The correct
configuration is setup in nvr.vim after plugin, but the fish
configuration, which does not properly split and wait for edits,
overrides it. Adding the correct config to fish causes some strange
errors and doesn't wait properly still, so instead of debugging that,
keep the configuration in nvr.vim, and guard the outside neovim variable
settings to only happen when not in a neovim session, so they don't
override the correct config inside.