Disable custom dict help mapping
It doesn't work the way I wanted, it has a wildcard can't be expanded
and if a split with dict is already open it causes another error.

Removing this no longer makes the content paged, which I want to fix,
but it is a lot faster and is still usable.