Experiment with putting mapleader back to default
The mapleader is normally \, but most people remap to ,. However, ,
already has a function. This removes that functionality, which is
actually useful, and causes a delay as Vim waits to see if another
character is coming. If I'm too slow, the overwritten function happens,
which might not be what I want.

My new keymappings make typing \ not that bad, and could be more
comfortable than ,.

Let's try it out.