Switch from submodules to packer
Git submodules are a pain in the ass to manager. I knew this before
using them. I thought the pain might be manageable. It kind of was, but I
don't want to deal with it. Packer, while another dependency, is
interesting, lua, and something I'll try for a while. It can manage the
native packages.