Refine nano email syntax highlighting
Remove markdown syntax highlighting, since it's not very accurate and gets it wrong a bunch. Generally not worth the time to get it right right now.

Replace the "all files" regex for Email with one that will never match, then tell Nano to use the Email group with a flag in my neomuttrc. Now my email syntax doesn't leak out to other files.

Right now, I only use nano to write emails, so this is fine (I go for nvim for heavier tasks).