Use jsonlint in neovim if available
jsonlint is better than the built-in python json linting, and someone
already wrote a compiler errorformat for it. I can get python's linting
to appear in the quick fix list, but it's not recognizing the line
number or column number, and I'm tired of trying to get it to work.