Fix Gup command
The async fugitive command to do a `git up` reports it's not in a work
tree when run. The function used to give it, and all the other working
fugitive commands the right directory gives the `.git` folder. While
this works for the other commands, it makes more sense to me to use the
work tree root, so switch all commands to use that.

Also, switch mappings to use the async versions of their commands, and
the non-deprecated version of the command.