Make init.lua global opts more global?
Calling `vim.o` vs `vim.opt` seems to have a functional difference,
outside returning a table for `opt`, in that `o`'s options seem to stick
like `set` when changing buffer types. Not sure exactly the behavior
difference, so this is one to keep an eye on.

Also, restore my old vimscript UrlView implementation. It's not that
elegant, using a temporary file, but it does work. Keep an eye here.

Finally, convert telescope config to lua.