Improve fzy search implementation
Use a temp file to save the output of the `fzy` command to edit, instead
of passing to `nvr`. This significantly improves the speed.

Make the function more flexible, taking arguments for the fzy input
command and the vim action to take.

Refactor to make the width almost as wide as the vim window, to better
see long lines.

Took inspiration from
https://git.icyphox.sh/dotfiles/tree/config/nvim/lua/fzy/shell.lua