Remove ipmotion plugin
This is a useful plugin, as it "improves" the paragraph movement options
to recognize a line with only whitespace characters. However, I'd rather
remove the plugin load and regex, in favor of the base system here,
since I have highlighting that shows a line with only whitespace as a
warning.