Fix markdown todo toggle mapping
The mapping function didn't properly check for a string containing the
`[x]` todo syntax, instead checking for the literal, meaning the toggle
would never work to turn "off" the checkbox.

Enhance the mapping by keeping the cursor position the same after the
job is done. Use the little used (for me) and hard to reach `p` register
to store the mark.