Remove git delta configuration and others
It's a nice program, but it's another piece of infrastructure in play,
takes longer to load, and I don't really need it.

Also, remove the default for the diff algorithm. I'll keep it off, try
out the default for a while and compare with others for larger diffsets
and all that.