Set EDITOR to ed
The EDITOR environment variable is a fallback for VISUAL: it only is
called if the terminal fails to produce the "advanced" rendering common
in visual editors (e.g., vim, emacs, etc), as opposed to line-oriented
editors, like ex/ed. This is a condition I have never come across, and
doubt it will come up, but to make it clear what the variable is for,
and to give it a useful value should the need ever arise, set it to ed.