Add gitsigns and comment.nvim
Replace commentary with comment.nvim. Tpope is a wizard, but I'll go
along and see what the treesitter integration provides practically.

Also add in the gitsigns plugin, as neovim's support for signs has
improved to the point where these types of plugins are no longer hacky
and dangerous (writing the buffer, etc).