Remove custom gx mapping
There is a native gx mapping now, and it is more robust. For example, it
works without error in terminal buffers, which don't have a file type
set, or buffers without a treesitter attached.

It only works in Markdown with the cursor positioned on the URL itself,
not the title, for example. I thought there was an open issue about this
but I can't find one.

Might be worth writing a small function to address.